+++ client/src/App.css
@@ -20,6 +20,7 @@
 * {
   font-family: "Anek Bangla", sans-serif;
   font-weight: 500;
+  margin: 0;
 }
 
 h1 {
