--- client/src/components/Profile/profile.css
@@ -1,107 +1,0 @@
-.profile-page {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: white;
-  padding: 20px;
-  background-color: #2c2c2c;
-}
-
-.profile-page h1 {
-  font-size: 1.7rem;
-  margin-bottom: 20px;
-  font-weight: 600;
-}
-
-.profile-page h2 {
-  font-size: 1.5rem;
-  font-weight: 300;
-}
-
-.profile-page p {
-  align-self: flex-start;
-}
-
-.user-info {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.modify-btn {
-  padding: 5px 20px;
-  border: none;
-  border-radius: 30px;
-  background-color: #6499ff;
-  color: white;
-  cursor: pointer;
-  margin-top: 10px;
-  font-size: large;
-}
-
-.edit-input {
-  font-size: 1rem;
-  padding: 5px;
-  border-radius: 10px;
-  border: 1px solid #ccc;
-  text-align: center;
-  margin-bottom: 10px;
-}
-
-.user-pic {
-  width: 70px;
-  height: 70px;
-  border-radius: 50%;
-  margin-bottom: 10px;
-  background-size: cover;
-  background-position: center;
-}
-
-.main-content {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 1240px;
-  padding: 0 20px;
-}
-
-.projects-collaborators-section {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin-bottom: 20px;
-  padding: 20px 0 20px 0;
-}
-
-@media screen and (min-width: 1241px) {
-  .main-content {
-    flex-direction: row;
-    justify-content: space-between;
-    padding: 0;
-  }
-  .projects-collaborators-section {
-    display: flex;
-    justify-content: space-around;
-    flex-direction: row;
-    width: 100%;
-  }
-}
-
-@media screen and (min-width: 600px) {
-  .profile-page {
-    padding: 10px;
-  }
-  .profile-page h1 {
-    font-size: 1.5rem;
-  }
-  .main-content {
-    padding: 0;
-  }
-  .task-manager-section,
-  .projects,
-  .collaborators {
-    width: 100%;
-    padding: 10px;
-  }
-}
