+++ client/src/pages/Home/Home.jsx
@@ -4,38 +4,16 @@ import TaskManager from "../../components/TaskManager/TaskManager";
 import Project from "../../components/Project/project";
 import Collaborater from "../../components/Collaborater/Collaborater";
 import Header from "../../components/Header/Header";
-import useProfile from "../../components/Profile/Profile";
+import profile from "../../assets/images/profile.jpg";
 
 function Home() {
-  const userId = 1;
-  const {
-    userName,
-    profilePic,
-    fetchUserName,
-    updateUserName,
-    updateProfilePic,
-  } = useProfile(userId);
-
   const [isEditing, setIsEditing] = useState(false);
+  const [userName, setUserName] = useState("John Doe");
   const [inputValue, setInputValue] = useState(userName);
+  const [profilePic, setProfilPic] = useState(profile);
+  const userId = 1;
   const inputRef = useRef(null);
   const fileInputRef = useRef(null);
-
-  useEffect(() => {
-    fetchUserName();
-  }, [fetchUserName]);
-
-  useEffect(() => {
-    if (isEditing && inputRef.current) {
-      inputRef.current.focus();
-    }
-  }, [isEditing]);
-
-  useEffect(() => {
-    if (!isEditing) {
-      updateUserName(inputValue);
-    }
-  }, [isEditing, inputValue, updateUserName]);
 
   const handleEditClick = () => {
     setIsEditing(true);
@@ -47,6 +25,7 @@ function Home() {
 
   const handleInputBlur = () => {
     setIsEditing(false);
+    setUserName(inputValue);
   };
 
   const handleInputKeyDown = (event) => {
@@ -54,17 +33,87 @@ function Home() {
       handleInputBlur();
     }
   };
-
-  const handleProfilePicClick = () => {
-    fileInputRef.current.click();
-  };
+ const handleProfilePicClick = () => {
+   fileInputRef.current.Click();
+ };
 
   const handleFileChange = async (event) => {
     const file = event.target.files[0];
     if (file) {
-      await updateProfilePic(file);
+      const formData = new FormData();
+      formData.append("profilePic", file);
+      formData.append("userId", userId);
+
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_API_URL}/api/users/update-profile-pic`,
+          {
+            method: "POST",
+            body: formData,
+          }
+        );
+
+        if (!response.ok) {
+          throw new Error("failed to update profile picture");
+        }
+        await response.json();
+        setProfilPic(URL.createObjectURL(file));
+      } catch (error) {
+        console.error(Error);
+      }
     }
   };
+
+  useEffect(() => {
+    const fetchUserName = async () => {
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_API_URL}/api/users/${userId}`
+        );
+        if (!response.ok) {
+          throw new Error("failed to fetch user data");
+        }
+        const data = await response.json();
+        setUserName(data.userName);
+        setInputValue(data.userName);
+        setProfilPic(data.profilePic || profile);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    fetchUserName();
+  }, [userId, userName]);
+
+  useEffect(() => {
+    if (isEditing && inputRef.current) {
+      inputRef.current.focus();
+    }
+  }, [isEditing]);
+
+  useEffect(() => {
+    const updateUserName = async () => {
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_API_URL}/api/users/update-name`,
+          {
+            method: "PUT",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ userId, newName: inputValue }),
+          }
+        );
+        if (!response.ok) {
+          throw new Error("Failed to update user name");
+        }
+        setUserName(inputValue);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+
+    if (!isEditing) {
+      updateUserName();
+    }
+  }, [userName, userId, isEditing, inputValue]);
 
   return (
     <>
@@ -86,7 +135,7 @@ function Home() {
             <h2>{`Bonjour, ${userName}`}</h2>
           )}
           <div
-            className="user-pic"
+           className="user-pic"
             role="button"
             onClick={handleProfilePicClick}
             onKeyDown={(e) => {
@@ -101,7 +150,7 @@ function Home() {
             ref={fileInputRef}
             style={{ display: "none" }}
             onChange={handleFileChange}
-          />
+          /> 
           <button
             type="button"
             className="modify-btn"
@@ -109,7 +158,7 @@ function Home() {
           >
             Modifier
           </button>
-        </div>
+       </div>
         <div id="main-content">
           <div className="task-manager-section">
             <TaskManager />
@@ -127,3 +176,4 @@ function Home() {
 }
 
 export default Home;
+
